refactor: create Repository test
QuestionRepositoryTest.java
UserRepositoryTest.java
